+++ integration_test/cat/cat_wallet_service_integration_test.dart
@@ -1,25 +1,23 @@
 // ignore_for_file: lines_longer_than_80_chars
 
 import 'package:chia_utils/chia_crypto_utils.dart';
-import 'package:chia_utils/src/api/simulator_utils.dart';
 import 'package:chia_utils/src/cat/puzzles/tails/delegated_tail/delegated_tail.clvm.hex.dart';
 import 'package:chia_utils/src/cat/puzzles/tails/genesis_by_coin_id/genesis_by_coin_id.clvm.hex.dart';
 import 'package:test/test.dart';
 
+import '../simulator/simulator_utils.dart';
+
 Future<void> main() async {
   const nTests = 3;
-
-  final simulatorUtils = SimulatorUtils();
-  try {
-    await simulatorUtils.checkIsRunning();
-  } catch(e) {
-    print(e);
+  
+  if(!(await SimulatorUtils.checkIfSimulatorIsRunning())) {
+    print(SimulatorUtils.simulatorNotRunningWarning);
     return;
   }
-
-  final simulatorHttpRpc = SimulatorHttpRpc(simulatorUtils.url,
-    certBytes: simulatorUtils.certBytes,
-    keyBytes: simulatorUtils.keyBytes,
+  
+  final simulatorHttpRpc = SimulatorHttpRpc(SimulatorUtils.simulatorUrl,
+    certBytes: SimulatorUtils.certBytes,
+    keyBytes: SimulatorUtils.keyBytes,
   );
   final fullNodeSimulator = SimulatorFullNodeInterface(simulatorHttpRpc);
 
@@ -40,11 +38,12 @@ Future<void> main() async {
   final senderWalletSet = keychain.unhardenedMap.values.first;
   final senderPuzzlehash = senderWalletSet.puzzlehash;
   final senderAddress = Address.fromPuzzlehash(senderPuzzlehash, catWalletService.blockchainNetwork.addressPrefix);
-
   for (var i = 0; i < nTests; i++) {
     await fullNodeSimulator.farmCoins(senderAddress);
   }
   await fullNodeSimulator.moveToNextBlock();
+
+  
 
   var senderStandardCoins = await fullNodeSimulator.getCoinsByPuzzleHashes([senderPuzzlehash]);
   final originCoin = senderStandardCoins[0];
@@ -54,7 +53,7 @@ Future<void> main() async {
   final assetId = Puzzlehash(curriedTail.hash());
   keychain.addOuterPuzzleHashesForAssetId(assetId);
   
-  final curriedGenesisByCoinIdPuzzle = genesisByCoinIdProgram.curry([Program.fromBytes(originCoin.id.toUint8List())]);
+  final curriedGenesisByCoinIdPuzzle = genesisByCoinIdProgram.curry([Program.fromBytes(originCoin.id)]);
   final tailSolution = Program.list([curriedGenesisByCoinIdPuzzle, Program.nil]);
 
   final signature = AugSchemeMPL.sign(senderWalletSet.childPrivateKey, curriedGenesisByCoinIdPuzzle.hash());
@@ -83,7 +82,12 @@ Future<void> main() async {
     // to avoid duplicate coins amounts must differ
     payments.add(Payment(990 + i, senderPuzzlehash));
   }
-  final sendBundle = catWalletService.createSpendBundle(payments, senderCatCoins, senderPuzzlehash, keychain);
+  final sendBundle = catWalletService.createSpendBundle(
+    payments: payments, 
+    catCoinsInput: senderCatCoins, 
+    changePuzzlehash: senderPuzzlehash, 
+    keychain: keychain,
+  );
 
   await fullNodeSimulator.pushTransaction(sendBundle);
   await fullNodeSimulator.moveToNextBlock();
@@ -109,7 +113,12 @@ Future<void> main() async {
     final amountToSend = (totalNateCoinValue * 0.8).round();
     final payment = Payment(amountToSend, receiverPuzzlehash);
 
-    final spendBundle = catWalletService.createSpendBundle([payment], catCoinsForThisTest, senderPuzzlehash, keychain);
+    final spendBundle = catWalletService.createSpendBundle(
+      payments: [payment], 
+      catCoinsInput: catCoinsForThisTest, 
+      changePuzzlehash: senderPuzzlehash, 
+      keychain: keychain,
+    );
     await fullNodeSimulator.pushTransaction(spendBundle);
     await fullNodeSimulator.moveToNextBlock();
 
@@ -137,7 +146,14 @@ Future<void> main() async {
     const fee = 1000;
     final payment = Payment(amountToSend, receiverPuzzlehash);
 
-    final spendBundle = catWalletService.createSpendBundle([payment], catCoinsForThisTest, senderPuzzlehash, keychain, fee: fee, standardCoinsForFee: standardCoinsForTest);
+    final spendBundle = catWalletService.createSpendBundle(
+      payments: [payment], 
+      catCoinsInput: catCoinsForThisTest, 
+      changePuzzlehash: senderPuzzlehash, 
+      keychain: keychain,
+      fee: fee, 
+      standardCoinsForFee: standardCoinsForTest,
+    );
     await fullNodeSimulator.pushTransaction(spendBundle);
     await fullNodeSimulator.moveToNextBlock();
 
@@ -182,7 +198,14 @@ Future<void> main() async {
 
     const fee = 1000;
 
-    final spendBundle = catWalletService.createSpendBundle(payments, catCoinsForThisTest, senderPuzzlehash, keychain, fee: fee, standardCoinsForFee: standardCoinsForTest);
+    final spendBundle = catWalletService.createSpendBundle(
+      payments: payments, 
+      catCoinsInput: catCoinsForThisTest, 
+      changePuzzlehash: senderPuzzlehash, 
+      keychain: keychain,
+      fee: fee, 
+      standardCoinsForFee: standardCoinsForTest,
+    );
     await fullNodeSimulator.pushTransaction(spendBundle);
     await fullNodeSimulator.moveToNextBlock();