+++ lib/src/api/models/responses/coin_records_response.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: lines_longer_than_80_chars
+
 import 'package:chia_utils/src/api/models/chia_models/chia_coin_record.dart';
 import 'package:chia_utils/src/api/models/responses/chia_base_response.dart';
 import 'package:meta/meta.dart';
@@ -34,4 +36,7 @@ class CoinRecordsResponse extends ChiaBaseResponse {
       error: chiaBaseResponse.error,
     );
   }
+
+  @override
+  String toString() => 'CoinRecordsResponse(coinRecords: $coinRecords, success: $success, error: $error)';
 }