+++ lib/src/core/exceptions/change_puzzlehash_needed_exception.dart
@@ -1,0 +1,8 @@
+// ignore_for_file: lines_longer_than_80_chars
+
+class ChangePuzzlehashNeededException implements Exception {
+  @override
+  String toString() {
+    return 'Change puzzle hash is required if this spend bundle will result and left over coins.';
+  }
+}