+++ lib/src/core/models/conditions/assert_coin_announcement_condition.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: lines_longer_than_80_chars
+
 import 'package:chia_utils/chia_crypto_utils.dart';
 import 'package:chia_utils/src/standard/exceptions/invalid_condition_cast_exception.dart';
 
@@ -30,7 +32,7 @@ class AssertCoinAnnouncementCondition implements Condition {
   Program get program {
     return Program.list([
       Program.fromInt(conditionCode),
-      Program.fromBytes(announcementId.toUint8List()),
+      Program.fromBytes(announcementId),
     ]);
   }
 
@@ -44,4 +46,7 @@ class AssertCoinAnnouncementCondition implements Condition {
     }
     return true;
   }
+
+  @override
+  String toString() => 'AssertCoinAnnouncementCondition(code: $conditionCode, coinId: $coinId, message: $message, morphBytes: $morphBytes)';
 }