+++ lib/src/core/models/spend_bundle.dart
@@ -1,8 +1,9 @@
 // ignore_for_file: lines_longer_than_80_chars
 
 import 'package:chia_utils/chia_crypto_utils.dart';
+import 'package:chia_utils/src/core/models/serializable.dart';
 
-class SpendBundle {
+class SpendBundle implements Serializable{
   List<CoinSpend> coinSpends;
   JacobianPoint? aggregatedSignature;
 
@@ -26,18 +27,35 @@ class SpendBundle {
       'coin_spends': coinSpends.map((e) => e.toJson()).toList(),
       'aggregated_signature': aggregatedSignature?.toHex(),
     };
+  SpendBundle.fromJson(Map<String, dynamic> json)
+    : coinSpends = (json['coin_solutions'] as Iterable).map((dynamic e) => CoinSpend.fromJson(e as Map<String, dynamic>)).toList(),
+      aggregatedSignature = JacobianPoint.fromHexG2(json['aggregated_signature'] as String); 
 
-  factory SpendBundle.aggregate(List<SpendBundle> spendBundles) {
+  SpendBundle operator +(SpendBundle other) {
     final signatures = <JacobianPoint>[];
-    var coinSpends = <CoinSpend>[];
-    for (final spendBundle in spendBundles) {
-      if(spendBundle.isSigned) {
-        signatures.add(spendBundle.aggregatedSignature!);
-      }
-      coinSpends += spendBundle.coinSpends;
+    if(aggregatedSignature != null) {
+      signatures.add(aggregatedSignature!);
     }
-    final aggregatedSignature = AugSchemeMPL.aggregate(signatures);
-    return SpendBundle(coinSpends: coinSpends, aggregatedSignature: aggregatedSignature);
+    if (other.aggregatedSignature != null) {
+      signatures.add(other.aggregatedSignature!);
+    }
+    return SpendBundle(
+      coinSpends: coinSpends + other.coinSpends,
+      aggregatedSignature: (signatures.isNotEmpty) ? AugSchemeMPL.aggregate(signatures) : null,
+    );
+  }
+
+  void addSignature(JacobianPoint signature) {
+    final signatures = <JacobianPoint>[signature];
+    if(aggregatedSignature != null) {
+      signatures.add(aggregatedSignature!);
+    }
+    aggregatedSignature = AugSchemeMPL.aggregate(signatures);
+  }
+
+  @override
+  Bytes toBytes() {
+    return serializeList(coinSpends) + Bytes(aggregatedSignature?.toBytes() ?? []);
   }
 
   void debug() {
@@ -49,4 +67,7 @@ class SpendBundle {
       print('result: ${spend.puzzleReveal.run(spend.solution).program}');
     }
   }
+
+  @override
+  String toString() => 'SpendBundle(coinSpends: $coinSpends, aggregatedSignature: $aggregatedSignature)';
 }