+++ lib/src/standard/exceptions/origin_id_not_in_coins_exception.dart
@@ -1,0 +1,6 @@
+class OriginIdNotInCoinsException implements Exception {
+  @override
+  String toString() {
+    return 'Origin id not in coins';
+  }
+}